pyproject.toml
[tool.poetry]
name = "splunk-opentelemetry"
version = "0.1.2"
description = "The Splunk distribution of OpenTelemetry Python Instrumentation provides a Python agent that automatically instruments your Python application to capture and report distributed traces to SignalFx APM."
authors = ["Splunk <splunk-oss@splunk.com>"]
license = "Apache-2.0"
packages = [
    { include = "splunk_otel" },
    { include = "splunk_otel/**/*.py" },
]
include = [
    'CHANGELOG.md',
    'README.md',
    'LICENSE'
]


[tool.poetry.scripts]
splk-py-trace = 'splunk_otel.cmd.trace:run'
splk-py-trace-bootstrap = 'splunk_otel.cmd.bootstrap:run'

[tool.poetry.dependencies]
python = "^3.5"
opentelemetry-api = "^0.13b0"
opentelemetry-sdk = "^0.13b0"
opentelemetry-instrumentation = "^0.13b0"
opentelemetry-exporter-zipkin = "^0.13b0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
mypy = "^0.782"
black = {version = "^20.8b1", python = "^3.6"}
isort = {version = "^5.5.3", python = "^3.6"}
pytest = "^6.1.0"

[tool.black]
line-length = 90
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

